# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zaps Backend API
  description: Backend API for the Zaps social payments platform on Stellar. Covers auth, profiles, feed, social interactions, and cross-chain token bridging.
  version: 1.0.0
  contact:
    name: Fracverse
servers:
  - url: http://localhost:8080
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /api/auth/verify
  responses:
...
    TooManyRequests:
      description: Too Many Requests - Rate limit exceeded; retry after a short delay
      content:
        text/plain:
          schema:
            type: string
          example: Too many requests, please try again later.
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
    ChallengeResponse:
      type: object
      required:
        - challenge
      properties:
        challenge:
          type: string
          description: UUID v4 challenge string to be signed by the client Stellar keypair
    VerifyRequest:
      type: object
      required:
        - address
        - signature
        - challenge
      properties:
        address:
          type: string
          description: Stellar G... public address
        signature:
          type: string
          description: Hex or base64-encoded Ed25519 signature of the challenge bytes
        challenge:
          type: string
          description: Challenge string previously obtained from GET /api/auth/challenge
    AuthResponse:
      type: object
      required:
        - token
        - username
      properties:
        token:
          type: string
          description: "JWT valid for 24 hours; include as Authorization: Bearer header"
        username:
          type: string
          description: Auto-generated username prefixed with u_ followed by chars 2-15 of the Stellar address
...
paths:
...
  /api/auth/challenge:
    get:
      tags:
        - Auth
      summary: Get a Stellar signing challenge
      operationId: getChallenge
      responses:
        '200':
          description: Challenge UUID to sign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeResponse'
              example:
                challenge: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/verify:
    post:
      tags:
        - Auth
      summary: Verify a Stellar signature and issue a JWT
      operationId: verifySignature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '200':
          description: Signature verified; JWT returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid signature format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Invalid signature format (must be hex or base64)
        '401':
          description: Signature verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Signature verification failed
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Internal database error
...